refactor: SearchBar에서 parmas를 내부에서 변경하도록 수정
